# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "name": "@aasumitro",
            "url": "https://aasumitro.id/",
            "email": "hello@aasumitro.id"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/aasumitro/posbe/blob/main/LICENSE"
        }
    },
    "paths": {
...
        "/v1/subcategories": {
            "get": {
                "description": "Get Subcategories List.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subcategories"
                ],
                "summary": "Subcategories List",
                "responses": {
                    "200": {
                        "description": "OK RESPOND",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SuccessRespond"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/domain.Subcategory"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorRespond"
                        }
                    },
                    "500": {
                        "description": "INTERNAL SERVER ERROR RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorRespond"
                        }
                    }
                }
            },
            "post": {
                "description": "Create new Subcategory.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subcategories"
                ],
                "summary": "Store Subcategory Data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "category_id",
                        "name": "category_id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "CREATED RESPOND",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SuccessRespond"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.Subcategory"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorRespond"
                        }
                    },
                    "422": {
                        "description": "UNPROCESSABLE ENTITY RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ValidationErrorRespond"
                        }
                    },
                    "500": {
                        "description": "INTERNAL SERVER ERROR RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorRespond"
                        }
                    }
                }
            }
        },
        "/v1/subcategories/{id}": {
            "put": {
                "description": "Update Subcategory Data by ID.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subcategories"
                ],
                "summary": "Update Subcategory Data",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "subcategory id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "category_id",
                        "name": "category_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CREATED RESPOND",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SuccessRespond"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.Subcategory"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "BAD REQUEST RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorRespond"
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorRespond"
                        }
                    },
                    "422": {
                        "description": "UNPROCESSABLE ENTITY RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ValidationErrorRespond"
                        }
                    },
                    "500": {
                        "description": "INTERNAL SERVER ERROR RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorRespond"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Subcategory Data by ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subcategories"
                ],
                "summary": "Delete Subcategory Data",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "subcategory id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "NO CONTENT RESPOND"
                    },
                    "400": {
                        "description": "BAD REQUEST RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorRespond"
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorRespond"
                        }
                    },
                    "500": {
                        "description": "INTERNAL SERVER ERROR RESPOND",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorRespond"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.Subcategory": {
            "type": "object",
            "required": [
                "category_id",
                "name"
            ],
            "properties": {
                "category_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "utils.ErrorRespond": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "utils.SuccessRespond": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "status": {
                    "type": "string"
                }
            }
        },
        "utils.ValidationErrorRespond": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "status": {
                    "type": "string"
                }
            }
        }
    }
}